Add comprehensive evaluee management and UI improvements
- Add new 'Manage Evaluees' page with full CRUD functionality
- Allow viewing all evaluees with detailed information and metrics
- Implement delete functionality with confirmation dialogs
- Add bulk operations including export and delete all
- Fix frequency_per_year field to support decimal values (e.g., 1.5 for "every 1.5 years")
- Enhance home page UI with better metrics display and quick actions
- Improve navigation flow with programmatic page switching
- Add enhanced database status display with comprehensive metrics
- Update FastAPI web interface to properly handle float frequency values

+++ web_app.py
@@ -147,7 +147,7 @@ async def add_service(
     table_name: str = Form(...),
     service_name: str = Form(...),
     unit_cost: Optional[float] = Form(None),
-    frequency_per_year: int = Form(...),
+    frequency_per_year: float = Form(...),
     inflation_rate: float = Form(...),
     service_type: Optional[str] = Form(None),  # "recurring" or "discrete" 
     start_year: Optional[int] = Form(None),
